seperated reducer,actions,types from the useEffect components & created use.types, use.actions & use.reducer files
