+++ src/componets/Use-Reducer/Use.reducer.jsx
@@ -1,0 +1,21 @@
+export const initialState = {
+  user: null,
+  searchQuery: "Bret",
+};
+
+export const userReducer = (state, action) => {
+  switch (action.type) {
+    case "SET_USER":
+      return {
+        ...state,
+        user: action.payload,
+      };
+    case "SET_SEARCH_QUERY":
+      return {
+        ...state,
+        searchQuery: action.payload,
+      };
+    default:
+      return state;
+  }
+};
