--- src/componets/Use-Reducer/useReducer.jsx
@@ -1,73 +1,0 @@
-import React, { useEffect, useReducer } from "react";
-import Card from "../Card/Card";
-
-const initialState = {
-  user: null,
-  searchQuery: "Bret"
-};
-
-const reducer = (state, action) => {
-  switch (action.type) {
-    case "SET_USER":
-      return {
-        ...state,
-        user: action.payload
-      };
-    case "SET_SEARCH_QUERY":
-      return {
-        ...state,
-        searchQuery: action.payload
-      };
-    default:
-      return state;
-  }
-};
-
-const setUser = user => ({
-  type: "SET_USER",
-  payload: user
-});
-
-const setSearchQuery = queryString => ({
-  type: "SET_SEARCH_QUERY",
-  payload: queryString
-});
-
-const UseEffect = () => {
-  const [state, dispatch] = useReducer(reducer, initialState);
-  const { user, searchQuery } = state;
-
-  useEffect(() => {
-    if (searchQuery.length > 0) {
-      const fetchFunc = async () => {
-        const response = await fetch(
-          `https://jsonplaceholder.typicode.com/users?username=${searchQuery}`
-        );
-        const data = await response.json();
-        dispatch(setUser(data[0]));
-      };
-      fetchFunc();
-    }
-  }, [searchQuery]);
-
-  return (
-    <Card>
-      <input
-        type="text"
-        value={searchQuery}
-        onChange={event => dispatch(setSearchQuery(event.target.value))}
-      />
-      {user ? (
-        <div>
-          <h3>{user.name}</h3>
-          <h3>{user.username}</h3>
-          <h3>{user.email}</h3>
-        </div>
-      ) : (
-        <p>No user found</p>
-      )}
-    </Card>
-  );
-};
-
-export default UseEffect;
